requirements.txt
defusedxml==0.7.1
fonttools==4.62.1
fpdf2==2.8.7
pillow==12.1.1
pywin32==311
watchdog==6.0.0
